Changement type tests et nouveaux tests
